refactor: renamed root folder
