+++ packages/atproto/lib/atproto.dart
@@ -3,12 +3,12 @@
 // modification, are permitted provided the conditions.
 
 export 'package:atproto/src/atproto.dart';
-export 'package:atproto/src/service/atproto_service.dart';
-export 'package:atproto/src/service/entities/record.dart';
-export 'package:atproto/src/service/repositories/repositories_service.dart';
-export 'package:atproto/src/service/sessions/current_session.dart';
-export 'package:atproto/src/service/sessions/session.dart';
-export 'package:atproto/src/service/sessions/sessions_service.dart';
+export 'package:atproto/src/atproto_service.dart';
+export 'package:atproto/src/entities/record.dart';
+export 'package:atproto/src/repositories/repositories_service.dart';
+export 'package:atproto/src/sessions/current_session.dart';
+export 'package:atproto/src/sessions/session.dart';
+export 'package:atproto/src/sessions/sessions_service.dart';
 export 'package:atproto_core/atproto_core.dart'
     show
         XRPCResponse,
