+++ packages/atproto/lib/src/entities/record.dart
