+++ packages/atproto/lib/src/sessions/current_session.freezed.dart
