+++ packages/atproto/lib/src/sessions/session.dart
